@@ -9,10 +9,10 @@ import os
 import numpy as np
 import traitlets
 from astropy import units as u
-from pixell import enmap, tilemap, fft
+from pixell import enmap, tilemap, fft, utils, bunch
 
 import toast
-from toast.traits import trait_docs, Unicode, Int, Instance, Bool, Float
+from toast.traits import trait_docs, Unicode, Int, Instance, Bool, Float, List
 from toast.ops import Operator
 from toast.utils import Logger, Environment, rate_from_times
 from toast.timing import function_timer, Timer
@@ -76,7 +76,17 @@ class MLMapmaker(Operator):
         help="The format is [from=](ra:dec|name),[to=(ra:dec|name)],[up=(ra:dec|name|system)]",
     )
 
-    comps = Unicode("T", help="Components (must be 'T', 'QU' or 'TQU')")
+    interpol = Unicode(
+        "nearest",
+        help="Either one or a comma-separated list of interpolation modes",
+    )
+
+    downsample = List(
+        [1],
+        help="Downsample TOD by these factors.",
+    )
+
+    comps = Unicode("TQU", help="Components (must be 'T', 'QU' or 'TQU')")
 
     Nmat = Instance(klass=mm.Nmat, allow_none=True, help="The noise matrix to use")
 
@@ -142,8 +152,20 @@ class MLMapmaker(Operator):
         help="If True, clear all observation detector data after accumulating",
     )
 
-    tiled = Bool(
+    checkpoint_interval = Int(
+        0,
+        help="If greater than zero, the CG solver will store its state and"
+        "restart from a checkpoint when available.",
+    )
+
+    skip_existing = Bool(
         False,
+        help="If True, the mapmaker will not write any map products that "
+        "already exist on disk.  See `checkpoint`."
+    )
+
+    tiled = Bool(
+        True,
         help="If True, the map will be represented as distributed tiles in memory. "
         "For large maps this is faster and more memory efficient, but for small "
         "maps it has some overhead due to extra communication."
@@ -163,12 +185,15 @@ class MLMapmaker(Operator):
     weather = Unicode("vacuum", help="Weather to assume when making maps")
     site    = Unicode("so",     help="Site to use when making maps")
 
-    maxiter = Int(500, help="Maximum number of CG iterations")
+    maxiter = List(
+        [500],
+        help="List of maximum number of CG iterations for each pass.",
+    )
 
     maxerr = Float(1e-6, help="Maximum error in the CG solver")
 
     truncate_tod = Bool(
-        False,
+        True,
         help="Truncate TOD to an easily factorizable length to ensure efficient FFT.",
     )
 
@@ -190,6 +215,13 @@ class MLMapmaker(Operator):
         check = proposal["value"]
         if check not in ["T", "QU", "TQU"]:
             raise traitlets.TraitError("Invalid comps (must be 'T', 'QU' or 'TQU')")
+        return check
+
+    @traitlets.validate("checkpoint_interval")
+    def _check_checkpoint_interval(self, proposal):
+        check = proposal["value"]
+        if check < 0:
+            raise traitlets.TraitError("Invalid checkpoint_interval. Must be non-negative.")
         return check
 
     @traitlets.validate("shared_flag_mask")
@@ -231,13 +263,6 @@ class MLMapmaker(Operator):
                 check = 1
         return check
 
-    @traitlets.validate("maxiter")
-    def _check_maxiter(self, proposal):
-        check = proposal["value"]
-        if check <= 0:
-            raise traitlets.TraitError("Maxiter should be greater than zero")
-        return check
-
     @traitlets.validate("maxerr")
     def _check_maxerr(self, proposal):
         check = proposal["value"]
@@ -262,12 +287,309 @@ class MLMapmaker(Operator):
         return check
 
     def __init__(self, **kwargs):
+        self.shape = None
+        self.wcs = None
+        self.recenter = None
+        self.signal_map = None
+        self.mapmaker = None
         super().__init__(**kwargs)
-        self._mapmaker = None
+
+    @function_timer
+    def setup_passes(self):
+        tmp = bunch.Bunch()
+        tmp.downsample = self.downsample
+        tmp.maxiter = self.maxiter
+        tmp.interpol = self.interpol.split(",")
+        # The entries may have different lengths. We use the max
+        # and then pad the others by repeating the last element.
+        # The final output will be a list of bunches
+        npass = max([len(tmp[key]) for key in tmp])
+        passes = []
+        for i in range(npass):
+            entry = bunch.Bunch()
+            for key in tmp:
+                entry[key] = tmp[key][min(i, len(tmp[key]) - 1)]
+            passes.append(entry)
+        return passes
+
+    @function_timer
+    def _load_noise_model(self, ob, npass, ipass, gcomm):
+        # Maybe load precomputed noise model
+        log = Logger.get()
+        if self.nmat_dir is None:
+            nmat_dir  = os.path.join(self.out_dir, "nmats")
+        else:
+            nmat_dir  = self.nmat_dir
+        if npass != 1:
+            nmat_dir += f"_pass{ipass + 1}"
+        nmat_file = os.path.join(nmat_dir, f"nmat_{ob.name}.hdf")
+        there = os.path.isfile(nmat_file)
+        if self.nmat_mode == "load" and not there:
+            raise RuntimeError(
+                f"Nmat mode is 'load' but {nmat_file} does not exist."
+            )
+        if self.nmat_mode == "load" or (self.nmat_mode == "cache" and there):
+            log.debug_rank(f"Loading noise model from '{nmat_file}'", comm=gcomm)
+            try:
+                nmat = mm.read_nmat(nmat_file)
+            except Exception as e:
+                if self.nmat_mode == "cache":
+                    log.info_rank(
+                        f"Failed to load noise model from '{nmat_file}'"
+                        f" : '{e}'. Will cache a new one",
+                        comm=gcomm,
+                    )
+                    nmat = None
+                else:
+                    msg = f"Failed to load noise model from '{nmat_file}' : {e}"
+                    raise RuntimeError(msg)
+        else:
+            nmat = None
+        return nmat, nmat_file
+
+    @function_timer
+    def _save_noise_model(self, mapmaker, nmat, nmat_file, gcomm):
+        # Maybe save the noise model we built (only if we actually built one rather than
+        # reading one in)
+        log = Logger.get()
+        if self.nmat_mode in ["save", "cache"] and nmat is None:
+            log.debug_rank(f"Writing noise model to '{nmat_file}'", comm=gcomm)
+            nmat_dir = os.path.dirname(nmat_file)
+            os.makedirs(nmat_dir, exist_ok=True)
+            mm.write_nmat(nmat_file, mapmaker.data[-1].nmat)
+        return
+
+    @function_timer
+    def _wrap_obs(self, ob, dets, passinfo):
+        """ Prepare data for the mapmaker """
+
+        # Get the focalplane for this observation
+        fp = ob.telescope.focalplane
+
+        # Get the sample rate from the data.  We also have nominal sample rates
+        # from the noise model and also from the focalplane.
+        # (rate, dt, dt_min, dt_max, dt_std) = rate_from_times(
+        #     ob.shared[self.times].data
+        # )
+
+        axdets = LabelAxis("dets", dets)
+        nsample = int(ob.n_local_samples)
+
+        axsamps = OffsetAxis(
+            "samps",
+            count=nsample,
+            offset=ob.local_index_offset,
+            origin_tag=ob.name,
+        )
+
+        # Convert the data view into a RangesMatrix
+        ranges = so3g.proj.ranges.RangesMatrix.zeros((len(dets), nsample))
+        if self.view is not None:
+             view_ranges = np.array(
+                 [[x.first, min(x.last, nsample) + 1] for x in ob.intervals[self.view]]
+             )
+             ranges += so3g.proj.ranges.Ranges.from_array(view_ranges, nsample)
+
+        # Convert the focalplane offsets into the expected form
+        det_to_row = {y["name"]: x for x, y in enumerate(fp.detector_data)}
+        det_quat = np.array([fp.detector_data["quat"][det_to_row[x]] for x in dets])
+        xi, eta, gamma = quat_to_xieta(det_quat)
+
+        axfp = AxisManager()
+        axfp.wrap("xi", xi, axis_map=[(0, axdets)])
+        axfp.wrap("eta", eta, axis_map=[(0, axdets)])
+        axfp.wrap("gamma", gamma, axis_map=[(0, axdets)])
+
+        # Convert Az/El quaternion of the detector back into
+        # angles from the simulation.
+        theta, phi, pa = toast.qarray.to_iso_angles(ob.shared[self.boresight])
+
+        # Azimuth is measured in the opposite direction from longitude
+        az = 2 * np.pi - phi
+        el = np.pi / 2 - theta
+        roll = pa
+
+        axbore = AxisManager()
+        axbore.wrap("az", az, axis_map=[(0, axsamps)])
+        axbore.wrap("el", el, axis_map=[(0, axsamps)])
+        axbore.wrap("roll", roll, axis_map=[(0, axsamps)])
+
+        axobs = AxisManager()
+        axobs.wrap("focal_plane", axfp)
+        axobs.wrap("timestamps", ob.shared[self.times], axis_map=[(0, axsamps)])
+        axobs.wrap(
+            "signal",
+            ob.detdata[self.det_data][dets, :],
+            axis_map=[(0, axdets), (1, axsamps)],
+        )
+        axobs.wrap("boresight", axbore)
+        axobs.wrap('flags', FlagManager.for_tod(axobs))
+        axobs.flags.wrap("glitch_flags", ranges, axis_map=[(0, axdets), (1, axsamps)])
+        axobs.wrap("weather", np.full(1, self.weather))
+        axobs.wrap("site",    np.full(1, "so"))
+
+        if self.truncate_tod:
+            # FFT-truncate for faster fft ops
+            axobs.restrict("samps", [0, fft.fft_len(axobs.samps.count)])
+
+        # MLMapmaker.add_obs will apply deslope
+        # if self.deslope:
+        #    utils.deslope(axobs.signal, w=5, inplace=True)
+
+        if self.downsample != 1:
+            axobs = mm.downsample_obs(axobs, passinfo.downsample)
+
+        # NOTE:  Expected contents look like:
+        # >>> tod
+        # AxisManager(signal[dets,samps], timestamps[samps], readout_filter_cal[dets],
+        # mce_filter_params[6], iir_params[3,5], flags*[samps], boresight*[samps],
+        # array_data*[dets], pointofs*[dets], focal_plane*[dets], abscal[dets],
+        # timeconst[dets], glitch_flags[dets,samps], source_flags[dets,samps],
+        # relcal[dets], dets:LabelAxis(63), samps:OffsetAxis(372680))
+        # >>> tod.focal_plane
+        # AxisManager(xi[dets], eta[dets], gamma[dets], dets:LabelAxis(63))
+        # >>> tod.boresight
+        # AxisManager(az[samps], el[samps], roll[samps], samps:OffsetAxis(372680))
+
+        return axobs
+
+    @function_timer
+    def _init_mapmaker(
+            self, mapmaker, signal_map, mapmaker_prev, x_prev, comm, gcomm, prefix,
+    ):
+        log = Logger.get()
+        timer = Timer()
+        timer.start()
+
+        mapmaker.prepare()
+
+        if self.tiled:
+            # Each group reports how many tiles they are using
+            geo_work = mapmaker.signals[1].geo_work
+            nactive = len(geo_work.active)
+            ntile = np.prod(geo_work.shape[-2:])
+            log.debug_rank(f"{nactive} / {ntile} tiles active", comm=gcomm)
+
+        if comm is not None:
+            comm.barrier()
+        log.info_rank(
+            f"MLMapmaker finished prepare in",
+            comm=comm,
+            timer=timer,
+        )
+
+        # This will need to be modified for more general cases where we don't solve for
+        # a sky map, or where we solve for multiple sky maps. The mapmaker itself supports it,
+        # the problem is the direct access to the rhs, div and idiv members
+        if self.write_rhs:
+            fname = f"{prefix}sky_rhs.fits"
+            if self.skip_existing and os.path.isfile(fname):
+                log.info_rank(f"Skipping existing rhs in {fname}", comm=comm)
+            else:
+                fname = signal_map.write(prefix, "rhs", signal_map.rhs)
+                log.info_rank(f"Wrote rhs to {fname}", comm=comm)
+
+        if self.write_div:
+            fname = f"{prefix}sky_div.fits"
+            if self.skip_existing and os.path.isfile(fname):
+                log.info_rank(f"Skipping existing div in {fname}", comm=comm)
+            else:
+                # FIXME : only writing the TT variance to avoid integer overflow in communication
+                fname = signal_map.write(prefix, "div", signal_map.div)
+                # fname = signal_map.write(prefix, "div", signal_map.div[0, 0])
+                log.info_rank(f"Wrote div to {fname}", comm=comm)
+
+        if self.write_hits:
+            fname = f"{prefix}sky_hits.fits"
+            if self.skip_existing and os.path.isfile(fname):
+                log.info_rank(f"Skipping existing div in {fname}", comm=comm)
+            else:
+                fname = signal_map.write(prefix, "hits", signal_map.hits)
+                log.info_rank(f"Wrote hits to {fname}", comm=comm)
+
+        mmul = tilemap.map_mul if self.tiled else enmap.map_mul
+        if self.write_bin:
+            fname = f"{prefix}sky_bin.fits"
+            if self.skip_existing and os.path.isfile(fname):
+                log.info_rank(f"Skipping existing bin in {fname}", comm=comm)
+            else:
+                fname = signal_map.write(
+                    prefix, "bin", mmul(signal_map.idiv, signal_map.rhs)
+                )
+                log.info_rank(f"Wrote bin to {fname}", comm=comm)
+
+        if comm is not None:
+            comm.barrier()
+        log.info_rank(f"MLMapmaker finished writing rhs, div, bin in", comm=comm, timer=timer)
+
+        # Set up initial condition
+
+        if x_prev is None:
+            x0 = None
+        else:
+            x0 = mapmaker.translate(mapmaker_prev, x_prev)
+
+        return x0
+
+    @function_timer
+    def _apply_mapmaker(self, mapmaker, x0, passinfo, prefix, comm):
+        log = Logger.get()
+        timer = Timer()
+        timer.start()
+        tstep = Timer()
+        tstep.start()
+
+        if self.checkpoint_interval > 0:
+            fname_checkpoint = f"{prefix}checkpoint.{comm.rank:04}.hdf"
+            there = os.path.isfile(fname_checkpoint)
+            if there:
+                log.info_rank(f"Checkpoint detected. Will start from previous solver state", comm=comm)
+        else:
+            fname_checkpoint = None
+
+        for step in mapmaker.solve(
+                maxiter=passinfo.maxiter,
+                maxerr=self.maxerr,
+                x0=x0,
+                fname_checkpoint=fname_checkpoint,
+                checkpoint_interval=self.checkpoint_interval,
+        ):
+            if self.write_iter_map < 1:
+                dump = False
+            else:
+                dump = step.i % self.write_iter_map == 0
+            msg = f"CG step {step.i:4d} {step.err:15.7e} write={dump}"
+            log.info_rank(f"MLMapmaker   {msg} ", comm=comm, timer=tstep)
+            if dump:
+                for signal, val in zip(mapmaker.signals, step.x):
+                    if signal.output:
+                        fname = signal.write(prefix, f"map{step.i:04}", val)
+                        log.info_rank(f"Wrote signal to {fname} in", comm=comm, timer=tstep)
+
+        log.info_rank(f"MLMapmaker finished solve in", comm=comm, timer=timer)
+
+        for signal, val in zip(mapmaker.signals, step.x):
+            if signal.output:
+                fname = signal.write(prefix, "map", val)
+                log.info_rank(f"Wrote {fname}", comm=comm)
+
+        if comm is not None:
+            comm.barrier()
+        log.info_rank(f"MLMapmaker wrote map in", comm=comm, timer=timer)
+
+        return mapmaker, step.x
 
     @function_timer
     def _exec(self, data, detectors=None, **kwargs):
         log = Logger.get()
+        timer = Timer()
+        comm = data.comm.comm_world
+        gcomm = data.comm.comm_group
+        timer.start()
+
+        if comm is None and self.tiled:
+            log.info("WARNING: Tiled mapmaking not supported without MPI.")
+            self.tiled = False
 
         for trait in ["area"]:
             value = getattr(self, trait)
@@ -277,268 +599,113 @@ class MLMapmaker(Operator):
                 )
 
         # nmat_type is guaranteed to be a valid Nmat class
-        self.Nmat = getattr(mm, self.nmat_type)()
+        noise_model = getattr(mm, self.nmat_type)()
 
-        comm = data.comm.comm_world
-        gcomm = data.comm.comm_group
+        shape, wcs = enmap.read_map_geometry(self.area)
 
-        if self._mapmaker is None:
-            # First call- create the mapmaker instance.
-            # Get the timestream dtype from the first observation
-            self._dtype_tod = data.obs[0].detdata[self.det_data].dtype
+        if self.center_at is None:
+            recenter = None
+        else:
+            recenter = mm.parse_recentering(self.center_at)
+        dtype_tod = np.float32
+        dtype_map = np.dtype(self.dtype_map)
 
-            self._shape, self._wcs = enmap.read_map_geometry(self.area)
+        prefix = os.path.join(self.out_dir, f"{self.name}_")
 
-            self._recenter = None
-            if self.center_at is not None:
-                self._recenter = mm.parse_recentering(self.center_at)
+        passes = self.setup_passes()
+        npass = len(passes)
+        mapmaker_prev = None
+        x_prev = None
 
-            dtype_tod    = np.float32
-            signal_cut   = mm.SignalCut(comm, dtype=dtype_tod)
-
-            signal_map   = mm.SignalMap(self._shape, self._wcs, comm, comps=self.comps,
-                               dtype=np.dtype(self.dtype_map), recenter=self._recenter, tiled=self.tiled)
-            signals      = [signal_cut, signal_map]
-            self._mapmaker = mm.MLMapmaker(signals, noise_model=self.Nmat, dtype=dtype_tod, verbose=self.verbose)
-            # Store this to be able to output rhs and div later
-            self._signal_map = signal_map
-
-        for ob in data.obs:
-            # Get the detectors we are using locally for this observation
-            dets = ob.select_local_detectors(detectors)
-            if len(dets) == 0:
-                # Nothing to do for this observation
-                continue
-
-            # Get the sample rate from the data.  We also have nominal sample rates
-            # from the noise model and also from the focalplane.
-            (rate, dt, dt_min, dt_max, dt_std) = rate_from_times(
-                ob.shared[self.times].data
+        for ipass, passinfo in enumerate(passes):
+            # The multipass mapmaking loop
+            log.info_rank(
+                f"Starting pass {ipass + 1}/{npass}, maxit={passinfo.maxiter} "
+                f"down={passinfo.downsample}, interp={passinfo.interpol}",
+                comm=comm,
             )
-
-            # Get the focalplane for this observation
-            fp = ob.telescope.focalplane
-
-            # Prepare data for the mapmaker.
-
-            axdets = LabelAxis("dets", dets)
-
-            nsample = int(ob.n_local_samples)
-            ind = slice(None)
-            ncut = nsample - fft.fft_len(nsample)
-            if ncut != 0:
-                if self.truncate_tod:
-                    log.info_rank(
-                        f"Truncating {ncut} / {nsample} samples ({100 * ncut / nsample:.3f}%) from "
-                        f"{ob.name} for better FFT performance.",
-                        comm=gcomm,
-                    )
-                    nsample -= ncut
-                    ind = slice(nsample)
-                else:
-                    log.warning_rank(
-                        f"{ob.name} length contains large prime factors.  "
-                        F"FFT performance may be degrared.  Recommend "
-                        f"truncating {ncut} / {nsample} samples ({100 * ncut / nsample:.3f}%).",
-                        comm=gcomm,
-                    )
-
-            axsamps = OffsetAxis(
-                "samps",
-                count=nsample,
-                offset=ob.local_index_offset,
-                origin_tag=ob.name,
-            )
-
-            # Convert the data view into a RangesMatrix
-            ranges = so3g.proj.ranges.RangesMatrix.zeros((len(dets), nsample))
-            if self.view is not None:
-                 view_ranges = np.array(
-                     [[x.first, min(x.last, nsample) + 1] for x in ob.intervals[self.view]]
-                 )
-                 ranges += so3g.proj.ranges.Ranges.from_array(view_ranges, nsample)
-
-            # Convert the focalplane offsets into the expected form
-            det_to_row = {y["name"]: x for x, y in enumerate(fp.detector_data)}
-            det_quat = np.array([fp.detector_data["quat"][det_to_row[x]] for x in dets])
-            xi, eta, gamma = quat_to_xieta(det_quat)
-
-            axfp = AxisManager()
-            axfp.wrap("xi", xi, axis_map=[(0, axdets)])
-            axfp.wrap("eta", eta, axis_map=[(0, axdets)])
-            axfp.wrap("gamma", gamma, axis_map=[(0, axdets)])
-
-            # Convert Az/El quaternion of the detector back into
-            # angles from the simulation.
-            theta, phi, pa = toast.qarray.to_iso_angles(ob.shared[self.boresight][ind])
-
-            # Azimuth is measured in the opposite direction from longitude
-            az = 2 * np.pi - phi
-            el = np.pi / 2 - theta
-            roll = pa
-
-            axbore = AxisManager()
-            axbore.wrap("az", az, axis_map=[(0, axsamps)])
-            axbore.wrap("el", el, axis_map=[(0, axsamps)])
-            axbore.wrap("roll", roll, axis_map=[(0, axsamps)])
-
-            axobs = AxisManager()
-            axobs.wrap("focal_plane", axfp)
-            axobs.wrap("timestamps", ob.shared[self.times][ind], axis_map=[(0, axsamps)])
-            axobs.wrap(
-                "signal",
-                ob.detdata[self.det_data][dets, ind],
-                axis_map=[(0, axdets), (1, axsamps)],
-            )
-            axobs.wrap("boresight", axbore)
-            axobs.wrap('flags', FlagManager.for_tod(axobs))
-            axobs.flags.wrap("glitch_flags", ranges, axis_map=[(0, axdets), (1, axsamps)])
-            axobs.wrap("weather", np.full(1, self.weather))
-            axobs.wrap("site",    np.full(1, "so"))
-
-            # NOTE:  Expected contents look like:
-            # >>> tod
-            # AxisManager(signal[dets,samps], timestamps[samps], readout_filter_cal[dets],
-            # mce_filter_params[6], iir_params[3,5], flags*[samps], boresight*[samps],
-            # array_data*[dets], pointofs*[dets], focal_plane*[dets], abscal[dets],
-            # timeconst[dets], glitch_flags[dets,samps], source_flags[dets,samps],
-            # relcal[dets], dets:LabelAxis(63), samps:OffsetAxis(372680))
-            # >>> tod.focal_plane
-            # AxisManager(xi[dets], eta[dets], gamma[dets], dets:LabelAxis(63))
-            # >>> tod.boresight
-            # AxisManager(az[samps], el[samps], roll[samps], samps:OffsetAxis(372680))
-
-            # Maybe load precomputed noise model
-            if self.nmat_dir is None:
-                nmat_dir  = os.path.join(self.out_dir, "nmats")
+            if npass == 1:
+                pass_prefix = prefix
             else:
-                nmat_dir  = self.nmat_dir
-            nmat_file = nmat_dir + "/nmat_%s.hdf" % ob.name
-            there = os.path.isfile(nmat_file)
-            if self.nmat_mode == "load" and not there:
-                raise RuntimeError(
-                    f"Nmat mode is 'load' but {nmat_file} does not exist."
+                pass_prefix = f"{prefix}pass{ipass + 1}_"
+
+            signal_cut = mm.SignalCut(comm, dtype=dtype_tod)
+            signal_map = mm.SignalMap(
+                shape,
+                wcs,
+                comm,
+                comps=self.comps,
+                dtype=dtype_map,
+                recenter=recenter,
+                tiled=self.tiled,
+                interpol=passinfo.interpol,
+            )
+            signals = [signal_cut, signal_map]
+            mapmaker = mm.MLMapmaker(
+                signals, noise_model=noise_model, dtype=dtype_tod, verbose=self.verbose
+            )
+
+            for ob in data.obs:
+                # Get the detectors we are using locally for this observation
+                dets = ob.select_local_detectors(detectors)
+                if len(dets) == 0:
+                    # Nothing to do for this observation
+                    continue
+
+                nmat, nmat_file = self._load_noise_model(ob, npass, ipass, gcomm)
+
+                axobs = self._wrap_obs(ob, dets, passinfo)
+                mapmaker.add_obs(
+                    ob.name,
+                    axobs,
+                    deslope=self.deslope,
+                    noise_model=nmat,
+                    signal_estimate=None,
                 )
-            if self.nmat_mode == "load" or (self.nmat_mode == "cache" and there):
-                log.info_rank(f"Loading noise model from '{nmat_file}'", comm=gcomm)
-                try:
-                    nmat = mm.read_nmat(nmat_file)
-                except Exception as e:
-                    if self.nmat_mode == "cache":
-                        log.info_rank(
-                            f"Failed to load noise model from '{nmat_file}'"
-                            f" : '{e}'. Will cache a new one",
-                            comm=gcomm,
-                        )
-                        nmat = None
-                    else:
-                        msg = f"Failed to load noise model from '{nmat_file}' : {e}"
-                        raise RuntimeError(msg)
-            else:
-                nmat = None
+                del axobs
 
-            self._mapmaker.add_obs(
-                ob.name, axobs, deslope=self.deslope, noise_model=nmat
+                self._save_noise_model(mapmaker, nmat, nmat_file, gcomm)
+
+                # Optionally delete the input detector data to save memory, if
+                # the calling code knows that no additional operators will be
+                # used afterwards.
+                if ipass == npass - 1 and self.purge_det_data:
+                    del ob.detdata[self.det_data]
+
+            if comm is not None:
+                comm.barrier()
+            log.info_rank(
+                f"MLMapmaker wrapped observations in",
+                comm=comm,
+                timer=timer,
             )
-            del axobs
 
-            # Maybe save the noise model we built (only if we actually built one rather than
-            # reading one in)
-            if self.nmat_mode in ["save", "cache"] and nmat is None:
-                log.info_rank(f"Writing noise model to '{nmat_file}'", comm=gcomm)
-                os.makedirs(nmat_dir, exist_ok=True)
-                mm.write_nmat(nmat_file, self._mapmaker.data[-1].nmat)
+            x0 = self._init_mapmaker(
+                mapmaker,
+                signal_map,
+                mapmaker_prev,
+                x_prev,
+                comm,
+                gcomm,
+                pass_prefix,
+            )
+            mapmaker_prev, x_prev = self._apply_mapmaker(
+                mapmaker, x0, passinfo, pass_prefix, comm
+            )
 
-            # Optionally delete the input detector data to save memory, if
-            # the calling code knows that no additional operators will be
-            # used afterwards.
-            if self.purge_det_data:
-                del ob.detdata[self.det_data]
+            # Save metadata, may get dropped later
+
+            self.shape = shape
+            self.wcs = wcs
+            self.recenter = recenter
+            self.signal_map = signal_map
+            self.mapmaker = mapmaker
 
         return
 
     @function_timer
     def _finalize(self, data, **kwargs):
-        # After multiple calls to exec, the finalize step will solve for the map.
-        log = Logger.get()
-        timer = Timer()
-        comm = data.comm.comm_world
-        gcomm = data.comm.comm_group
-        timer.start()
-
-        self._mapmaker.prepare()
-
-        if self.tiled:
-            geo_work = self._mapmaker.signals[1].geo_work
-            nactive = len(geo_work.active)
-            ntile = np.prod(geo_work.shape[-2:])
-            log.info_rank(f"{nactive} / {ntile} tiles active", comm=gcomm)
-
-        log.info_rank(
-            f"MLMapmaker finished prepare in",
-            comm=comm,
-            timer=timer,
-        )
-
-        prefix = os.path.join(self.out_dir, f"{self.name}_")
-
-        # This will need to be modified for more general cases where we don't solve for
-        # a sky map, or where we solve for multiple sky maps. The mapmaker itself supports it,
-        # the problem is the direct access to the rhs, div and idiv members
-        if self.write_rhs:
-            fname = self._signal_map.write(prefix, "rhs", self._signal_map.rhs)
-            log.info_rank(f"Wrote rhs to {fname}", comm=comm)
-
-        if self.write_div:
-            #self._signal_map.write(prefix, "div", self._signal_map.div)
-            # FIXME : only writing the TT variance to avoid integer overflow in communication
-            fname = self._signal_map.write(prefix, "div", self._signal_map.div[0, 0])
-            log.info_rank(f"Wrote div to {fname}", comm=comm)
-
-        if self.write_hits:
-            fname = self._signal_map.write(prefix, "hits", self._signal_map.hits)
-            log.info_rank(f"Wrote hits to {fname}", comm=comm)
-
-        mmul = tilemap.map_mul if self.tiled else enmap.map_mul
-        if self.write_bin:
-            fname = self._signal_map.write(
-                prefix, "bin", mmul(self._signal_map.idiv, self._signal_map.rhs)
-            )
-            log.info_rank(f"Wrote bin to {fname}", comm=comm)
-
-        if comm is not None:
-            comm.barrier()
-        log.info_rank(f"MLMapmaker finished writing rhs, div, bin in", comm=comm, timer=timer)
-
-        tstep = Timer()
-        tstep.start()
-
-        for step in self._mapmaker.solve(maxiter=self.maxiter, maxerr=self.maxerr):
-            if self.write_iter_map < 1:
-                dump = False
-            else:
-                dump = step.i % self.write_iter_map == 0
-            dstr = ""
-            if dump:
-                dstr = "(write)"
-            msg = f"CG step {step.i:4d} {step.err:15.7e} {dstr}"
-            log.info_rank(f"MLMapmaker   {msg} ", comm=comm, timer=tstep)
-            if dump:
-                for signal, val in zip(self._mapmaker.signals, step.x):
-                    if signal.output:
-                        fname = signal.write(prefix, "map%04d" % step.i, val)
-                        log.info_rank(f"Wrote signal to {fname}", comm=comm)
-
-        log.info_rank(f"MLMapmaker finished solve in", comm=comm, timer=timer)
-
-        for signal, val in zip(self._mapmaker.signals, step.x):
-            if signal.output:
-                fname = signal.write(prefix, "map", val)
-                log.info_rank(f"Wrote {fname}", comm=comm)
-
-        if comm is not None:
-            comm.barrier()
-        log.info_rank(f"MLMapmaker wrote map in", comm=comm, timer=timer)
+        pass
 
     def _requires(self):
         req = {
@@ -565,51 +732,3 @@ class MLMapmaker(Operator):
 
     def _accelerators(self):
         return list()
-
-
-# class NmatToast(mm.Nmat):
-#     """Noise matrix class that uses a TOAST noise model.
-
-#     This takes an existing TOAST noise model and uses it for a MLMapmaker compatible
-#     noise matrix.
-
-#     Args:
-#         model (toast.Noise):  The toast noise model.
-#         det_order (dict):  The mapping from detector order in the AxisManager
-#             to name in the Noise object.
-
-#     """
-#     def __init__(self, model, n_sample, det_order):
-#         self.model = model
-#         self.det_order = det_order
-#         self.n_sample = n_sample
-
-#         # Compute the radix-2 FFT length to use
-#         self.fftlen = 2
-#         while self.fftlen <= self.n_sample:
-#             self.fftlen *= 2
-#         self.npsd = self.fftlen // 2 + 1
-
-#         # Compute the time domain offset that centers our data within the
-#         # buffer
-#         self.padded_start = (self.fftlen - self.n_sample) // 2
-
-#         # Compute the common frequency values
-#         self.nyquist = model.freq(model.keys[0])[-1].to_value(u.Hz)
-#         self.rate = 2 * self.nyquist
-#         self.freqs = np.fft.rfftfreq(self.fftlen, 1 / self.rate))
-
-#         # Interpolate the PSDs to desired spacing and store for later
-#         # application.
-
-#     def build(self, tod, **kwargs):
-#         """Build method is a no-op, we do all set up in the constructor."""
-#         return self
-
-#     def apply(self, tod, inplace=False):
-#         """Apply our noise filter to the TOD.
-
-#         We use our pre-built Fourier domain kernels.
-
-#         """
-#         return tod
