@@ -80,7 +80,12 @@ def set_book_rebind(imprint, book, update_level2=False):
 
     if op.exists(book_dir):
         print(f"Removing all files from {book_dir}")
-        shutil.rmtree(book_dir)
+        if os.path.isfile(book_dir):
+            os.remove(book_dir)
+        elif os.path.isdir(book_dir):
+            shutil.rmtree(book_dir)
+        else:
+            print("How is this not a file or directory")
     else: 
         print(f"Found no files in {book_dir} to remove")
 
@@ -160,6 +165,21 @@ def block_set_rebind(imprint, update_level2=False):
         imprint.logger.info(f"Setting book {book.bid} for rebinding")
         set_book_rebind(imprint, book, update_level2=update_level2)    
 
+def block_fix_bad_timing(imprint):
+    """Run through and try rebinding all books with bad timing"""
+    failed_books = imprint.get_failed_books()
+    fix_list = []
+    for book in failed_book:
+        if "TimingSystemOff" in book.message:
+            fix_list.append(book)
+    for book in fix_list:
+        imprint.logger.info(f"Setting book {book.bid} for rebinding")
+        set_book_rebind(imprint, book)
+        imprint.logger.info(
+            f"Binding book {book.bid} while accepting bad timing"
+        )
+        imprint.bind_book(book, allow_bad_timing=True)
+
 def get_timecode_final(imprint, time_code, type='all'):
     """Check if all required entries in the g3tsmurf database are present for
     smurf or stray book regisitration.
@@ -186,16 +206,17 @@ def get_timecode_final(imprint, time_code, type='all'):
     g3session, SMURF = imprint.get_g3tsmurf_session(return_archive=True)
     session = imprint.get_session()
 
+    # this is another place I was reminded sqlite does not accept
+    # numpy int32s or numpy int64s
+    time_code = int(time_code)
+
     servers = SMURF.finalize["servers"]
     meta_agents = [s["smurf-suprsync"] for s in servers]
     files_agents = [s["timestream-suprsync"] for s in servers]
 
-    meta_query = or_(*[TimeCodes.agent == a for a in meta_agents])
-    files_query = or_(*[TimeCodes.agent == a for a in files_agents])
-
     tcm = g3session.query(TimeCodes.agent).filter(
         TimeCodes.timecode==time_code,
-        meta_query,
+        TimeCodes.agent.in_(meta_agents),
         TimeCodes.suprsync_type == SupRsyncType.META.value,
     ).distinct().all()
 
@@ -209,8 +230,8 @@ def get_timecode_final(imprint, time_code, type='all'):
         return False, 1
 
     tcf = g3session.query(TimeCodes.agent).filter(
-        TimeCodes.timecode==time_code,
-        files_query,
+        TimeCodes.timecode == time_code,
+        TimeCodes.agent.in_(files_agents),
         TimeCodes.suprsync_type == SupRsyncType.FILES.value,
     ).distinct().all()
     
@@ -244,8 +265,10 @@ def set_timecode_final(imprint, time_code):
     """
 
     g3session, SMURF = imprint.get_g3tsmurf_session(return_archive=True)
-
     servers = SMURF.finalize["servers"]
+    # this is another place I was reminded sqlite does not accept
+    # numpy int32s or numpy int64s
+    time_code = int(time_code)
     
     for server in servers:
         tcf = g3session.query(TimeCodes).filter(
@@ -261,6 +284,7 @@ def set_timecode_final(imprint, time_code):
                 agent=server["timestream-suprsync"],
             )
             g3session.add(tcf)
+            g3session.commit()
 
         tcm = g3session.query(TimeCodes).filter(
             TimeCodes.timecode==time_code,
@@ -274,5 +298,5 @@ def set_timecode_final(imprint, time_code):
                 timecode=time_code,
                 agent=server["smurf-suprsync"],
             )
-            g3session.add(tcm)    
-    g3session.commit()
+            g3session.add(tcm)     
+            g3session.commit()
